refactored code to rescale DCF/CCF by purity
