+++ src/decifer/generator.py
@@ -5,7 +5,6 @@ author: Simone Zaccaria
 date: 2020-06-04
 """
 
-
 import sys, os
 import argparse
 import warnings
